Time: 0 ms (100%), Space: 97.4 MB (24.2%) - LeetHub

+++ 0121-best-time-to-buy-and-sell-stock/0121-best-time-to-buy-and-sell-stock.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int maxProfit(vector<int>& prices) {
         int minPrice = prices[0], best = 0;
-        for (int i=1; i<prices.size(); ++i){
+        for (int i=1; i<(int)prices.size(); ++i){
             minPrice = min(minPrice, prices[i]);
             best     = max(best, prices[i] - minPrice);
         }
